Port optionstr to Rust
