+++ Filelist
@@ -109,9 +109,8 @@ SRC_ALL =	\
 		src/ops.c \
 		src/option.c \
 		src/option.h \
-		src/optiondefs.h \
-		src/optionstr.c \
-		src/profiler.c \
+                src/optiondefs.h \
+                src/profiler.c \
                 src/regexp.h \
                 src/register.c \
 		src/scriptfile.c \
@@ -278,7 +277,6 @@ SRC_ALL =	\
 		src/proto/normal.pro \
 		src/proto/ops.pro \
 		src/proto/option.pro \
-		src/proto/optionstr.pro \
 		src/proto/popupmenu.pro \
 		src/proto/popupwin.pro \
 		src/proto/profiler.pro \
