+++ rust_option/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 crate-type = ["staticlib"]
 
 [dependencies]
+rust_optionstr = { path = "../rust_optionstr" }
 
 [build-dependencies]
 bindgen = "0.69"
