+++ rust_optionstr/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "rust_optionstr"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+crate-type = ["staticlib", "rlib"]
+
+[dependencies]
