+++ src/Make_mvc.mak
@@ -736,7 +736,6 @@ OBJ = \
 	$(OUTDIR)\normal.obj \
 	$(OUTDIR)\ops.obj \
 	$(OUTDIR)\option.obj \
-	$(OUTDIR)\optionstr.obj \
 	$(OUTDIR)\os_mswin.obj \
 	$(OUTDIR)\os_win32.obj \
 	$(OUTDIR)\pathdef.obj \
@@ -1631,7 +1630,6 @@ $(OUTDIR)/normal.obj: $(OUTDIR) normal.c $(INCL) nv_cmdidxs.h nv_cmds.h
 
 $(OUTDIR)/option.obj: $(OUTDIR) option.c $(INCL) optiondefs.h
 
-$(OUTDIR)/optionstr.obj: $(OUTDIR) optionstr.c $(INCL)
 
 $(OUTDIR)/ops.obj: $(OUTDIR) ops.c $(INCL)
 
@@ -1892,7 +1890,6 @@ proto.h: \
 	proto/normal.pro \
 	proto/ops.pro \
 	proto/option.pro \
-	proto/optionstr.pro \
 	proto/os_mswin.pro \
 	proto/winclip.pro \
 	proto/os_win32.pro \
