+++ src/proto.h
@@ -176,7 +176,6 @@ void mbyte_im_set_active(int active_arg);
 # include "normal.pro"
 # include "ops.pro"
 # include "option.pro"
-# include "optionstr.pro"
 # include "popupmenu.pro"
 # if defined(FEAT_PROFILE) || defined(FEAT_RELTIME)
 #  include "profiler.pro"
